# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CinePrime",
    "description": "CinePrime is a movie booking platform that allows users to book movies and watch them in the cinema.",
    "version": "1.0.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/movie-showtime/add": {
      "post": {
        "tags": [
          "Movies"
        ],
        "summary": "Add Showtime",
        "description": "Add Showtime",
        "operationId": "addShowtime",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "movie_id": {
                    "type": "number",
                    "example": 49
                  },
                  "price": {
                    "type": "number",
                    "example": 20
                  },
                  "show_time": {
                    "type": "string",
                    "example": "2024-05-01 10:00:00"
                  }
                }
              },
              "examples": {
                "Add Showtime": {
                  "value": {
                    "movie_id": 49,
                    "price": 20,
                    "show_time": "2024-05-01 10:00:00"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}